// src/codegen/models/DeleteProjectTaskStatusResponse.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type DeleteProjectTaskStatusResponse = {
  /**
   * Indicates whether the operation was successful or not
   */
  success: boolean;
  /**
   * Deleted project task status ID
   */
  data: Record<string, any>;
  /**
   * If the operation was not successful, holds the deatils of errors
   */
  errors: Array<string>;
};