// src/codegen/models/GetProjectResponse.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { Project } from './Project';

export type GetProjectResponse = {
  /**
   * Indicates whether the operation was successful or not
   */
  success: boolean;
  /**
   * The fetched project
   */
  data: Project;
  /**
   * If the operation was not successful, holds the deatils of errors
   */
  errors: Array<string>;
};